+++ eslint.config.ts
@@ -26,23 +26,34 @@ export default eslintConfig({
     {
       files: TEST_DIR_GLOB,
       rules: {
-        'sort-imports': 0,
+        curly: 0,
         'import/order': 0,
-        'unused-imports/no-unused-imports': 0,
-        'import/no-duplicates': 0,
-        'import/newline-after-import': 0,
         'no-unused-vars': 0,
-        '@typescript-eslint/no-unused-vars': 0,
-        '@stylistic/padding-line-between-statements': 0,
-        '@typescript-eslint/no-non-null-assertion': 0,
+        'prefer-template': 0,
+        'sort-imports': 0,
+
+        'unused-imports/no-unused-imports': 0,
+
+        'import/newline-after-import': 0,
+        'import/no-duplicates': 0,
+
         '@typescript-eslint/consistent-type-imports': 0,
         '@typescript-eslint/no-explicit-any': 0,
+        '@typescript-eslint/no-non-null-assertion': 0,
+        '@typescript-eslint/no-shadow': 0,
+        '@typescript-eslint/no-unused-vars': 0,
+
+        '@stylistic/padding-line-between-statements': 0,
         '@stylistic/quotes': 0,
-        curly: 0,
+
         'jsdoc/lines-before-block': 0,
-        'prefer-template': 0,
-        'astro/sort-attributes': 0,
+
         'astro/prefer-class-list-directive': 0,
+        'astro/sort-attributes': 0,
+
+        'svelte/button-has-type': 0,
+        'svelte/prefer-style-directive': 0,
+        'svelte/sort-attributes': 0,
       },
     },
   ],